Split part of xDSLOpt in an xDSLTool base class.
