+++ tests/xdsl_opt/test_xdsl_opt.py
@@ -6,8 +6,9 @@ import pytest
 from xdsl.dialects import builtin
 from xdsl.ir import MLContext
 from xdsl.passes import ModulePass
+from xdsl.tools.xdsl_tool import get_all_dialects
 from xdsl.utils.exceptions import DiagnosticException
-from xdsl.xdsl_opt_main import get_all_dialects, get_all_passes, xDSLOptMain
+from xdsl.xdsl_opt_main import get_all_passes, xDSLOptMain
 
 
 def test_dialects_and_passes():
