+++ xdsl/tools/irdl_to_pyrdl.py
@@ -7,7 +7,7 @@ from xdsl.dialects.irdl import DialectOp
 from xdsl.dialects.irdl.irdl_to_pyrdl import convert_dialect
 from xdsl.ir import MLContext
 from xdsl.parser import Parser
-from xdsl.xdsl_opt_main import get_all_dialects
+from xdsl.tools.xdsl_tool import get_all_dialects
 
 
 def main():
